OBPIH-6155 Move importPicklist method to PicklistController
