+++ grails-app/controllers/org/pih/warehouse/picklist/PicklistController.groovy
@@ -12,13 +12,16 @@ package org.pih.warehouse.picklist
 import grails.converters.JSON
 import grails.gorm.transactions.Transactional
 import org.pih.warehouse.api.PickPageItem
+import org.pih.warehouse.api.StockMovement
 import org.pih.warehouse.core.Constants
 import org.pih.warehouse.core.DocumentService
 import org.pih.warehouse.core.Location
 import org.pih.warehouse.data.DataService
+import org.pih.warehouse.importer.ImportDataCommand
 import org.pih.warehouse.inventory.StockMovementService
 import org.pih.warehouse.order.Order
 import org.pih.warehouse.requisition.Requisition
+import org.springframework.web.multipart.MultipartFile
 
 @Transactional
 class PicklistController {
@@ -173,5 +176,32 @@ class PicklistController {
             default:
                 throw new IllegalFormatException("Unable to determine the proper rendering format for request for format ${format}")
         }
+    }
+
+    def importPickListItems(ImportDataCommand command) {
+        StockMovement stockMovement = stockMovementService.getStockMovement(params.id)
+        List<PickPageItem> pickPageItems = stockMovementService.getPickPageItems(params.id, null, null, false)
+
+        MultipartFile importFile = command.importFile
+        if (importFile.isEmpty()) {
+            throw new IllegalArgumentException("File cannot be empty")
+        }
+
+        String csv = new String(importFile.bytes)
+        List<Map> importedLines = stockMovementService.parsePickCsvTemplateImport(csv)
+
+        stockMovementService.validatePicklistListImport(importedLines, pickPageItems)
+
+        List errors = importedLines*.errors.withIndex().collect { errors, index ->
+            errors.collect { "Row ${1}: ${it}" }
+        }.flatten()
+
+        stockMovementService.importPicklistTemplate(importedLines, stockMovement, pickPageItems)
+
+        if (!errors.isEmpty()) {
+            render([message: "Data imported with errors", errors: errors] as JSON)
+        }
+
+        render([message: "Data imported successfully"] as JSON)
     }
 }
